2025_Rendu_de_projet_LANDREIN_BUREAU/Code/requirements.txt
flask
nibabel
numpy
scikit-image
open3D
numpy-stl
setuptools